+++ src/main/java/com/hackybear/hungry_scan_core/controller/cms/ThemeController.java
@@ -34,7 +34,7 @@ public class ThemeController {
 
     @PostMapping("/show")
     @PreAuthorize("hasAnyRole('MANAGER', 'ADMIN')")
-    public ResponseEntity<Map<String, Object>> show(@RequestBody Integer id) {
+    public ResponseEntity<Map<String, Object>> show(@RequestBody Long id) {
         return responseHelper.getResponseEntity(id, themeService::findById);
     }
 
@@ -46,7 +46,7 @@ public class ThemeController {
 
     @PostMapping("/set-active")
     @PreAuthorize("hasAnyRole('MANAGER', 'ADMIN')")
-    public ResponseEntity<Map<String, Object>> setActive(@RequestBody Integer id) {
+    public ResponseEntity<Map<String, Object>> setActive(@RequestBody Long id) {
         return responseHelper.buildResponse(id, themeService::setActive);
     }
