+++ src/main/java/com/hackybear/hungry_scan_core/repository/AllergenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AllergenRepository extends JpaRepository<Allergen, Integer> {
+public interface AllergenRepository extends JpaRepository<Allergen, Long> {
 }
