+++ src/main/java/com/hackybear/hungry_scan_core/repository/ZoneRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ZoneRepository extends JpaRepository<Zone, Integer> {
+public interface ZoneRepository extends JpaRepository<Zone, Long> {
     List<Zone> findAllByOrderByDisplayOrder();
 }
